backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MT5 Integration
MetaTrader5==5.0.45

# AI/LLM
openai==1.3.7
anthropic==0.7.7
langchain==0.1.0
langchain-openai==0.0.2

# Data Analysis
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
pandas-ta==0.3.14b0

# News and Data Sources
newsapi-python==0.2.7
alpha-vantage==2.3.1
yfinance==0.2.32
requests==2.31.0
beautifulsoup4==4.12.2

# Database
motor==3.3.2
pymongo==4.6.0
sqlalchemy==2.0.23

# WebSocket and Real-time
python-socketio==5.10.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
python-dateutil==2.8.2

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
